install: Extract autoconf-archive install to separate script
Under some circumstances it is needed to 'hand install' the m4
scripts. This was done in the rhel install script, but sometimes
it is really useful to be able to do that by hand, so extract the
relevant commands into their own script (to allow hand running),
and get the rhel script to invoke that new script at the appropriate
time.

Fixes: #875

Signed-off-by: Graham Whaley <graham.whaley@intel.com>